[Fix] Rename and Refactor(Sign -> User)
