+++ src/main/java/walkbook/server/controller/UserController.java
@@ -2,36 +2,34 @@ package walkbook.server.controller;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
-import walkbook.server.dto.sign.TokenResponse;
+import walkbook.server.dto.post.PostResponse;
+import walkbook.server.dto.user.TokenResponse;
+import walkbook.server.dto.user.UserRequest;
 import walkbook.server.dto.user.UserResponse;
-import walkbook.server.dto.sign.SignInRequest;
-import walkbook.server.dto.sign.SignUpRequest;
 import walkbook.server.dto.CommonResponse;
 import walkbook.server.dto.SingleResponse;
 import walkbook.server.service.SignService;
 import walkbook.server.service.UserService;
 import walkbook.server.service.response.ResponseService;
 
-import javax.validation.Valid;
-
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/api/user")
-public class SignController {
+public class UserController {
     private final ResponseService responseService;
     private final SignService signService;
     private final UserService userService;
 
     @PostMapping("/signin")
-    public TokenResponse createAuthenticationToken(@Valid @RequestBody SignInRequest signInRequest) {
-        String username = signInRequest.getUsername();
-        final String token = signService.signin(signInRequest);
+    public TokenResponse createAuthenticationToken(@RequestBody UserRequest userRequest) {
+        String username = userRequest.getUsername();
+        final String token = signService.signin(userRequest);
         return responseService.getTokenResult(new UserResponse(userService.findByUsername(username)), token);
     }
 
     @PostMapping("/signup")
-    public CommonResponse registerUser(@Valid @RequestBody SignUpRequest signUpRequest) {
-        Long userId = signService.signup(signUpRequest);
+    public CommonResponse registerUser(@RequestBody UserRequest userRequest) {
+        Long userId = signService.signup(userRequest);
         return responseService.getSingleResult(new UserResponse(userService.findById(userId)));
     }
 
@@ -39,4 +37,14 @@ public class SignController {
     public SingleResponse<UserResponse> getUserInfo(@PathVariable Long userId) {
         return responseService.getSingleResult(new UserResponse(userService.findById(userId)));
     }
+
+//    @GetMapping("/{userId}/liked-post")
+//    public SingleResponse<PostResponse> getUserLikePosts(@PathVariable Long userId){
+//        return responseService.getListResult();
+//    }
+//
+//    @GetMapping("/{userId}/my-post")
+//    public SingleResponse<PostResponse> getUserPosts(@PathVariable Long userId){
+//        return responseService.getListResult();
+//    }
 }
