--- src/main/java/walkbook/server/dto/sign/SignInRequest.java
@@ -1,22 +1,0 @@
-package walkbook.server.dto.sign;
-
-import lombok.*;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import walkbook.server.domain.User;
-
-@Getter
-@Setter
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class SignInRequest {
-    private String username;
-    private String password;
-
-    public User toUser(PasswordEncoder passwordEncoder) {
-        return User.builder()
-                .username(username)
-                .password(passwordEncoder.encode(password))
-                .build();
-    }
-}
