+++ src/main/java/walkbook/server/dto/user/UserRequest.java
@@ -4,6 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import walkbook.server.domain.Gender;
 import walkbook.server.domain.User;
 
@@ -12,17 +13,18 @@ import walkbook.server.domain.User;
 @AllArgsConstructor
 @NoArgsConstructor
 public class UserRequest {
-
     private String username;
+    private String password;
     private String nickname;
     private Gender gender;
     private String age;
     private String location;
     private String introduction;
 
-    public User toEntity() {
+    public User toEntity(PasswordEncoder passwordEncoder) {
         return User.builder()
                 .username(username)
+                .password(passwordEncoder.encode(password))
                 .nickname(nickname)
                 .gender(gender)
                 .age(age)
