+++ src/main/java/walkbook/server/repository/PostRepository.java
@@ -7,6 +7,5 @@ import walkbook.server.domain.Post;
 
 public interface PostRepository extends JpaRepository<Post, Long> {
     Page<Post> findAll(Pageable pageable);
-
     Page<Post> findByTitleContainingOrDescriptionContaining(Pageable pageable, String title, String description);
 }
